requirements.txt
appdirs==1.4.3
appnope==0.1.0
beautifulsoup4==4.6.0
bleach==2.0.0
boto==2.48.0
boto3==1.4.4
botocore==1.5.88
cached-property==1.3.0
certifi==2017.4.17
chardet==3.0.4
Collectfast==0.5.2
coreapi==2.3.1
coreschema==0.0.4
cycler==0.10.0
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.3
django-allauth==0.32.0
django-appconf==1.0.2
django-classy-tags==0.8.0
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-easy-maps==0.9.3
django-extensions==1.8.1
django-htmlmin==0.10.0
django-jchart==0.4.1
django-pg-utils==0.1.5
django-redis==4.8.0
django-redis-cache==1.7.1
django-rest-swagger==2.1.2
django-s3-folder-storage==0.5
django-storages==1.6.3
django-tables2==1.10.0
django-url-filter==0.3.3
djangorestframework==3.6.3
docutils==0.13.1
durable-rules==0.33.114
entrypoints==0.2.3
enum34==1.1.6
geopy==1.11.0
gunicorn==19.7.1
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
oauthlib==2.0.2
openapi-codec==1.3.2
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.3
ptyprocess==0.5.2
Pycco==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
pystache==0.5.4
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
rcssmin==1.0.6
reactive==0.1
redis==2.10.6
requests==2.18.2
requests-oauthlib==0.8.0
rjsmin==1.0.12
s3transfer==0.1.10
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
smartypants==2.0.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
